Shrink array element types to match actual value ranges
particleStack[]: int -> unsigned char (indices 0-41 only)
whichKernel[]: int -> unsigned char (matches kernel's own type,
  _KERNEL_* values are all tiny)
rasterBleed.c f[]: int -> unsigned char (small frame-count index)
sin_cos[32]: int -> short (Q8 values, -256..256)
line85[]: short int -> signed char (values -1..127)
drawScreen.c sh[]: unsigned int -> unsigned char (shift amounts 0-7)
R_PENTOMINO[5][2]: int -> unsigned char (values 0-2)
score.c base[], drawScreen.c arenas[]: unsigned int -> unsigned short
  (RAM buffer offsets, well under 65536)
sound.c multiplier[]: int -> unsigned short (max value 0x400)

Also removed playerAnimation.h's `extern const int AnimationIndex[]`
-- never defined or referenced anywhere, dead declaration.

Left debug[CH_MAX], Attribute[]/isActive[] (28 flag bits), pwr[],
_pitchTable[], pix85[], input_counter/target[] untouched -- each
genuinely needs its current width.

+++ main/particle.h
@@ -27,7 +27,7 @@ struct Particle {
 };
 
 extern struct Particle particle[PARTICLE_COUNT];
-extern const int sin_cos[32];
+extern const short sin_cos[32];
 extern int weapon;
 
 void initParticles();
